chore(auto-edit): integrate model response metadata with analytics logger
