+++ vscode/src/autoedits/analytics-logger/index.ts
@@ -1,1 +1,2 @@
 export * from './analytics-logger'
+export * from './types'
